backend/requirements.txt
fastapi
uvicorn
numpy
torch
pandas
python-multipart
python-dotenv
requests